Move create_price_finder to price_finding module
